+++ src/libacars/util.h
@@ -10,7 +10,7 @@
 #include <stdlib.h>		// free()
 
 typedef struct {
-	uint8_t id;
+	int id;
 	void *val;
 } la_dict;
 
@@ -21,6 +21,6 @@ void *la_xrealloc(void *ptr, size_t size, const char *file, const int line, cons
 #define LA_XREALLOC(ptr, size) la_xrealloc((ptr), (size), __FILE__, __LINE__, __func__)
 #define LA_XFREE(ptr) do { free(ptr); ptr = NULL; } while(0)
 
-void *la_dict_search(const la_dict *list, uint8_t id);
+void *la_dict_search(const la_dict *list, int id);
 size_t la_slurp_hexstring(char *string, uint8_t **buf);
 #endif // !LA_UTIL_H
